fix(): fix merge preview routes
