+++ packages/brick-kit/src/core/Router.spec.ts
@@ -310,80 +310,6 @@ describe("Router", () => {
     expect(kernel.prefetchDepsOfStoryboard).toBeCalled();
   });
 
-  it("it should work-MergePreviewRouter", async () => {
-    const router = new Router(kernel);
-    const result = router.MergePreviewRouter([
-      {
-        path: "${APP.homepage}/home",
-      },
-      {
-        path: "${APP.homepage}/_dev_only_/template-preview/test",
-      },
-      {
-        path: "${APP.homepage}/test1",
-      },
-      {
-        path: "${APP.homepage}/_dev_only_/form-preview/test",
-      },
-      {
-        path: "${APP.homepage}/test2",
-      },
-      {
-        path: "${APP.homepage}/_dev_only_/snippet-preview/test",
-      },
-    ]);
-    expect(result).toEqual([
-      {
-        path: "${APP.homepage}/home",
-      },
-      {
-        path: "${APP.homepage}/_dev_only_/template-preview/test",
-      },
-      {
-        bricks: [
-          {
-            brick: "span",
-          },
-        ],
-        exact: true,
-        menu: false,
-        path: "${APP.homepage}/_dev_only_/template-preview/:templateId",
-      },
-      {
-        path: "${APP.homepage}/test1",
-      },
-      {
-        path: "${APP.homepage}/_dev_only_/form-preview/test",
-      },
-      {
-        bricks: [
-          {
-            brick: "span",
-          },
-        ],
-        exact: true,
-        menu: false,
-        path: "${APP.homepage}/_dev_only_/form-preview/:FormId",
-      },
-      {
-        path: "${APP.homepage}/test2",
-      },
-      {
-        path: "${APP.homepage}/_dev_only_/snippet-preview/test",
-      },
-      {
-        bricks: [
-          {
-            brick: "span",
-          },
-        ],
-        exact: true,
-        menu: false,
-        path: "${APP.homepage}/_dev_only_/snippet-preview/:snippetId",
-      },
-    ]);
-  });
-
   it("should redirect to login page if not logged in.", async () => {
     __setMatchedStoryboard({
       app: {
