+++ packages/brick-kit/src/core/Router.ts
@@ -1,12 +1,10 @@
 import { locationsAreEqual, createPath, Action, Location } from "history";
-import { findLastIndex, uniqueId } from "lodash";
+import { uniqueId } from "lodash";
 import type {
-  BaseRouteConf,
   LayoutType,
   PluginHistoryState,
   PluginLocation,
   PluginRuntimeContext,
-  RouteConf,
   RuntimeMisc,
 } from "@next-core/brick-types";
 import {
@@ -59,6 +57,7 @@ import {
   getStandaloneInstalledApps,
   preFetchStandaloneInstalledApps,
 } from "../internal/getStandaloneInstalledApps";
+import { mergePreviewRoutes } from "../internal/mergePreviewRoutes";
 
 export class Router {
   private defaultCollapsed = false;
@@ -328,7 +327,7 @@ export class Router {
         },
       };
       try {
-        const mergedRoutes = this.MergePreviewRouter(storyboard.routes);
+        const mergedRoutes = mergePreviewRoutes(storyboard.routes);
         await locationContext.mountRoutes(
           mergedRoutes,
           undefined,
@@ -578,57 +577,6 @@ export class Router {
     };
   }
 
-  MergePreviewRouter(router: RouteConf[]): RouteConf[] {
-    let mergedRoutes = router;
-    const specificTemplatePreviewIndex = findLastIndex(mergedRoutes, (route) =>
-      route.path.startsWith("${APP.homepage}/_dev_only_/template-preview/")
-    );
-    if (specificTemplatePreviewIndex > -1) {
-      mergedRoutes = [
-        ...mergedRoutes.slice(0, specificTemplatePreviewIndex + 1),
-        {
-          path: "${APP.homepage}/_dev_only_/template-preview/:templateId",
-          bricks: [{ brick: "span" }],
-          menu: false,
-          exact: true,
-        } as RouteConf,
-        ...mergedRoutes.slice(specificTemplatePreviewIndex + 1),
-      ];
-    }
-
-    const specificSnippetPreviewIndex = findLastIndex(mergedRoutes, (route) =>
-      route.path.startsWith("${APP.homepage}/_dev_only_/snippet-preview/")
-    );
-    if (specificSnippetPreviewIndex > -1) {
-      mergedRoutes = [
-        ...mergedRoutes.slice(0, specificSnippetPreviewIndex + 1),
-        {
-          path: "${APP.homepage}/_dev_only_/snippet-preview/:snippetId",
-          bricks: [{ brick: "span" }],
-          menu: false,
-          exact: true,
-        } as RouteConf,
-        ...mergedRoutes.slice(specificSnippetPreviewIndex + 1),
-      ];
-    }
-
-    const specificFormPreviewIndex = findLastIndex(mergedRoutes, (route) =>
-      route.path.startsWith("${APP.homepage}/_dev_only_/form-preview/")
-    );
-    if (specificFormPreviewIndex > -1) {
-      mergedRoutes = [
-        ...mergedRoutes.slice(0, specificFormPreviewIndex + 1),
-        {
-          path: "${APP.homepage}/_dev_only_/form-preview/:FormId",
-          bricks: [{ brick: "span" }],
-          menu: false,
-          exact: true,
-        } as RouteConf,
-        ...mergedRoutes.slice(specificFormPreviewIndex + 1),
-      ];
-    }
-    return mergedRoutes;
-  }
   /* istanbul ignore next */
   getNavConfig(): NavConfig {
     return this.navConfig;
